Pull out some cpp files.
